googler: Switch from http.client to urllib.request
urllib.request is higher level than http.client and does almost
everything we need automatically (redirection, for instance). It even
recognizes users' proxy settings in the https_proxy env var (when we
don't set a proxy explicitly). Most importantly, it saves a hell lot of
code and maintenance.

The only downside is that `Connection:close' is hardcoded in each
request* so we can't keep the connection alive. I doubt that will cause
any noticeable delay on any reasonable network in a country or region
where Google is usable.

Note that I'm building an URL opener just to be forward-looking. We
could also have used urllib.request.urlopen for the moment.

* See https://docs.python.org/3/library/urllib.request.html:

  > urllib.request module uses HTTP/1.1 and includes Connection:close
  > header in its HTTP requests.